user resource: add initial tests
These tests are not enabled yet. We can enable when the resource
code lands.

Note: We need to handle the case where a users's roles ordering can
change.  The flattened version of the terraform state available in
acceptance tests can make that trickier than it should be. I've
tried to implement some helper functions that may be reusable (or
reusable in part) for other cases where we need to match an unordered
set.