Extend Role resource with permissions support (#108)
Adds the ability for a user to set `permissions` for the role they are
creating by setting a valid JSON string with the `permissions`
attribute.

The provider will correctly compare between API permissions, which are
partially computed, and those set by the user.
If permissions are set by the user, then those permissions will be
stored in the statefile.
If no permissions are set by the user, then the API-computed permissions
will be stored in the statefile.
On import of a Role resource, the existing API permissions will be
imported and stored in the statefile.

Adds tests to verify correct behaviour of permissions and import of
permissions.

New features:
- Add `permissions` support to Role resource.
- Also add `multitenant_locked`, `landing_url` fields.
- Remove prior restriction on `role_type` field: Now supports `user` and
`account` role types.
- Update generated schema.

Testing:
-  Update existing tests to use added attribute fields.
- Remove usage of `.md.tmpl` files in non-documentation example tests.
- Add tests for computed permissions, user-set default access
permissions, user-set feature permissions via string and via
`jsonencode()`
- Add test helper function for getting a Role.