Add AnyToValue convert function
We can use this when working with dynamic attributes.

The general pattern (pseudo code) will be:

start with a `plan.MyDynamicAttribute`

Get the types `t` of the underlying firstObject, then:

```
var x map[string]any
AnyToValue(ctx, x, t)
```

A future function will allow conversion in the
reverse direction.

```
secondObject := ValueToAny(ctx, x)
```

Comparing objects will be possible with:

```
secondObject.Equal(firstObject)
```

When AnyToValue encounters fields that are not
present in the reference types it wil skip them.
(We may need additional ways to handle such
fields in the future, but I expect 'skip' will
be the most frequent pattern.)