Add ValueToAny convert function
This is the inverse of the previously added AnyToValue.

In (simplified) pseudocode:

```
x := AnyToValue(X)
Y := ValueToAny(x)
Y == X is true
```

It should be possible, for example to use ValueToAny
when using a DynamicAttribute, to create the contents
of a suitable POST request.

Similarly, it should be possible to use AnyToValue when
using a DynamicAttribute, to map the contents of
a GET's json body to suitable form for state file output.