Add tests for using all permissions to create user and account roles (#140)
Adds acceptance tests for creating a user role and account role using
all the permissions.
VDI pool permissions are not tested for now as the OpenAPI spec for VDI
Pools is currently inaccurate.

For creating resources which are dependencies, we use a mix of the old
and new Terraform provider.