Switch to using jsontypes.Normalized for role permissions (#144)
Changes `permissions` to use `jsontypes.Normalized` instead of
`types.String`. This means that any valid JSON string provided to the
`permission` field will be stored with `json.encode()` in state. This
gives us some benefits, mainly that we don't have to worry about sorting
JSON to store to state, and that a valid JSON string literal is now
stored as encoded JSON to state. This gives the provider more resilience
against the user who wants to use a literal string in the `permissions`
field.

These changes are more or less a drop-in replacement for what was there.
All of our tests still pass so this should be an easy change.
`jsontypes.Normalized` wraps `types.String` so doing this gives us the
additional conveniences of the new type.