Refactor permissions block to use Terraform Plugin Framework types (#159)
This PR switches to using Terraform plugin framework types for the Role
permissions block.

Key changes in this PR:
- Generated schema has been updated.
- Permissions block and all permissions it contains are `optional` and
act as overrides. We only track them in state if the user has set them
in the state file.
- The subset algorithm has been changed. We now use a simpler method
involving `slices.IndexFunc` to determine whether permissions in state
are still a valid subset of those in API state.
- Lots of boilerplate additions to handle conversions between generated
SDK structs and Terraform Plugin Framework types.
- Added an implementation for the `ValidateConfig` method for the Role
resource. This prevents a user from setting cloud permissions on a user
role and group permissions on an account role.