Perform role type validation on default group/cloud access levels (#165)
We've agreed to enforce additional validation of these fields on the
provider side.
This cleans up importing of a role, as now in Terraform state the
incompatible `default access` level will be stored as null.
Morpheus UI suggests that the unsupported `default access` level for the
given role type should not be accessible to a user.