Refactor role acceptance tests to use strongly typed permissions (#166)
Refactors the Role acceptance tests to use the strongly typed
permissions.

This involves updating existing tests which are still relevant with the
strongly typed permissions, and removing tests which are no longer
necessary - such as those relating to JSON.

Furthermore, the acceptance tests have been refactored to use the
`acctest.RandomWithPrefix()` function to generate random names for the
resources created.

We're left with the following tests so far, which covers all fields of
the role resource:

```
TestAccMorpheusRoleRequiredAttrsOk
TestAccMorpheusRoleAllAttrsOk
TestAccMorpheusRoleExampleOk
TestAccMorpheusRolePermissionsDefaultAccessPermissionsOk
TestAccMorpheusRoleExampleLegacyProviderOk
TestAccMorpheusRoleAllPermissionsUserRoleOk
TestAccMorpheusRoleAllPermissionsAccountRoleOk
```