network resource: add update
For the Dynamic Attribute `config` block, we assume that
changes to the block are HTTP `PUT`-able, and thus supported
by update.

The `config` block entries cannot be fetched by HTTP GET, so
we assume that a successful PUT (200) will have updated the values
successfully.

This may not be true for all network types, but it seems a
pragmatic first pass for network support.