network resource: add import tests (#188)
We create an import specific test file: import_test.go
This is consistent with the pattern of creating operation
specific files (import.go, etc).

We follow the same pattern as `user` and other resources,
we use create to create a resource and then a second
test phase to import and delete that resource.