Return empty slice when Set/List is null or unknown
`FromSetType` and `FromListType` are intended to be used to convert a terraform set/list to a slice to be used with the API. The functions can currently cause errors if the value is null or unknown when we try to convert the elements.
We don't want these functions to return an error, but instead return an empty slice when set/list contains no elements.