Add Role Update functionality (#201)
This PR adds functionality for the Update method to the Role resource.

- Implements Update method
- Adds helper functions for setting permissions in Update and for
resetting permissions in advance of the final Update operation.
- Tweaks the feature permission subset/compute algorithm in Read to set
the `Access` field so that we can detect drift in state.

Update is complicated by a couple of things:
- For feature permissions, the Morpheus API returns a fixed length,
computed list on a GET.
- For non-feature permissions, the Morpheus API returns only permissions
for those resources whose access level has been set to a non-`default`
value (e.g., `full` or `none`). That means if you haven't changed the
access level for a resource from `default`, it won't display any
permissions for that resource in the GET.

This means that to make the provider work as expected, we need to reset
ALL permissions access levels with each Update() before sending the PUT
request containing our planned permission levels.
 
This makes it so a user can remove a non-feature permission setting from
their Terraform config without causing a drift in state. Without
resetting the permissions on each update, if a user was to remove a
non-feature fine-grained permission setting from their config after
setting it in a prior apply, it would still be set in the API but no
longer set in Terraform state. Doing this means we can treat the
permissions settings in the provider as overrides which revert back to
default values if removed from a config.