Add UseStateForNullOrUnknown Bool plan modifier (#217)
This is required to iron out some problems introduced by #215.

If we run `terraform plan` as part of an update, the `multitenant` and
`multitenant_locked` fields will be marked as `known after apply` when
we modify a field that isn't one of those two.

The built-in `UseStateForUnknown` modifier only covers the case where
the planned value is unknown, not null.

Since these fields can be persisted in state as either `true/false` or
`null`, with an apply on a `tenant` role maintaining a literal `null`
value since we don't want to track it in state, we'll need a slightly
modified version of `UseStateForUnknown`.