Revert "Add UseStateForNullOrUnknown Bool plan modifier" (#221)
Reverts HPE/terraform-provider-hpe#217

Reverting this PR as further investigation which led to #220 showed that
I could not get the desired behaviour from the attributes using this
plan modifier, so there's no need for this anymore. Furthermore,
`optional` behaviour on an attribute does not behave as expected using
this plan modifier.