Modify User acceptance tests to use generated test names (#229)
This change is made to help prevent test failures when we have dangling
resources.

As with the other resources' acceptance tests, In some cases of test
failure we may still have dangling resources, but this will make it so
on a subsequent run a different name will be generated for the user
resources to be created so that we do not have clashes with existing
resources from a prior test run.

To note, the test `TestAccMorpheusUserExample` should be updated at some
point to use the `testhelpers.RenderExample()` function and an
`example.tf.tmpl` file should be used in place of the static
`resource.tf` which is currently being used for testing and
documentation. For now we're using a string substitution to get the
generated name into the config read from the file as a string.