Add error checking to the instance polling
Currently we are not checking the final error of the polling code
This can cause problems if during provisioning an instance goes into one of the error states as the error will not be handled.

Updates the code to add the error checking of the polling code and adds a more user friendly message.