Support outputting a debug env file
Newer versions of `terraform-plugin-framework` support outputting the debug configuration into a file.

This allows us remove the manual steps when we want to run the provider with a debugger.

Update the go.mod with latest version of `terraform-plugin-framework` and update `main.go` to support writing out the debug config file.