Add utility to compare plan and returned Attribute stanzas
In this PR we:
- create a new internal/utils directory and package
- add a new function CheckPlanAttributeAgainstAPIAttribute and
  unit-tests for it

This function will compare an arbitrary Plan Attribute stanza against an
arbitrary corresponding Attribute stanza returned by the API.  At the
moment it only checks that the keys are the same.  It can be extended in
future to check the types of the values and the actual values
themselves.