network resource: add sleep to import test (#324)
We are seeing issues with stray network resources
remaining after acceptance test runs.

The working theory is that this is caused by
a race condition: if a delete occurs very soon
after a create, the original network may get
deleted, but the usual second network that's
created ("synced") in a second cloud does
not get deleted.

Adding a sleep to confirm whether or not this is a race.