Update attribute_compare to work with ObjectValues that are empty
When importing certain datastores, the Config block returned by the API
has extra information over that specified by the user.  However in
import the plan is empty and thus the keys used in the ObjectValue
cannot be determined from the Attributes.  To get around these
situations we extract the AttributeTypes when ObjectValue is empty and
use this to construct the map of keys.  This means that the keys which
can be set by the user are extracted from the Config block returned by
the API.