WIP: Instance updating and resizing (#316)
Adding an implementation of instance updating and resizing.
This is currently not finished and requires some further work.

The code here _essentially_ works, but will cause problems in certain
scenarios.
To make sure that Terraform understands how the elements in volumes
array
have changed during an update I've changed the volume name to be
required
and have changed volumes from a set to a list. This also more accurately
reflects the API.

The plan is to sort volumes by name in a plan modifier, this will allow
terraform to
better keep track of the elements. (currently not implemented)
I've also needed to add custom `RequiresReplace` statements, because the
default one pretty much always forces a replacement.

Currently the code doesn't detect volume removal nicely, we might need
to add
a read-only "index" attribute to the volume elements which is added to
the state
and compared during the update to detect missing elements.

The `schema_gen.go` file was manually edited here to speed up
development,
the changes will need to be ported to internal configs.

---------

Co-authored-by: Jason Quigley <jason.quigley@hpe.com>
Co-authored-by: Jason Quigley <138007593+jasonquigleyhpecom@users.noreply.github.com>