Use state values for Visibility and Active on Update
If they aren't set in the config.  Vivek found an issue where the PUT
would return a 500 on the Update call during Create if Visiblity and
Active hadn't been set in the config.  On debugging the issue is that
if those values aren't set in the config then Morpheus sets them for
you. We can't use a PlanModifier (UseStateForUnknown) since that
requires that the State exists and is written out, which isn't the case
on Create.  So we add code to use the State values for both - which
are set to those returned by the API - if the values are Unknown in
the Plan, i.e. not set by the user.

Note that When we Create we need to run an Update directly afterward
to set the resourcePermissions and tenants.