Switch to using custom legacy SDK for sdkv2 provider (#378)
This PR switches the client used by the sdkv2 provider in
`internal/sdkv2/morpheus/config.go` to our custom legacy SDK.

The `config.Client()` method has been updated to support this.

Additionally:
- Adds a constructor `NewLegacyClient()` which uses our custom
RoundTrippers for token and creds authentication.
- Adds a test for `NewLegacyClient` which mocks the authentication flow
of Morpheus with the legacy SDK. This is to ensure the RoundTrippers
work as expected with our configuration of the client.