Add a builder function for reporting type assertion errors (#386)
Adds a builder function to help with constructing errors for reporting
failed type assertions.

This is useful as we will aim to make the legacy provider code more type
safe as we port it.

We can use this to return a diagnostic wrapping an error whenever we hit
a failed assertion and avoid the provider panicking.