Implement in-place updates for the Image resource
Adds in-place updating for most attributes in the Image resource.

`url` and `config_azure` will continue to require a replace as those fields may require extra logic to ensure
correct updating of the underlying image.

`description` requires a replace due to the API not supporting description updates at the moment.