Having default case of StringToBool be null
Having the default case of StringToBool conversion return null instead of false
to reflect the absence of a value more accurately and for consistency with the
other functions of the API.

For example, BoolToType returns nil for nil input.
```
func BoolToType(b *bool) types.Bool {
	if b == nil {
		return types.BoolNull()
	}

	return types.BoolValue(*b)
}
```

This is somewhat relevant to the policy datasource as currently - bools from
the API are represented as null while those which have been converted from strings
are represented as false - which could. be misleading.
```
config_lifecycle            = {
          + account_integration_id               = null
          + lifecycle_age                        = null
            // StringToBool booleans
          + lifecycle_allow_extend               = false
          + lifecycle_auto_renew                 = false

          + lifecycle_extensions_before_approval = null

          // BoolToType boolean
          + lifecycle_hide_fixed                 = null

          + lifecycle_message                    = null
          + lifecycle_notify                     = null
          + lifecycle_renewal                    = null
```