Add BMaaS instance example and tidy-up instance docs
We've added another instance example, this time one for BMaaS.  We've
also tidied up the instance docs and elevated the examples one level up
as regards headings so that they appear as links in the page index.

Finally we've change the "config" stanzas to include the following:
- noAgent = false
- createUser = true

Since those settings are likely to be more common amongst customers.