Fix instance intefaces order in state-file
There was a bug in the code use to process the list of interfaces in the
server part of a container which meant that the order of interface
"names" (e.g. eth0, eth1 etc) would be changed.  This meant that an
instance plan would report a replace since the order of interfaces in
the state-file would be changed.  This PR fixes that bug.

We've added an example HCL with two networks to the docs.