Instance network interface and update changes
In this PR we make changes to the instance network interfaces Schema.
We add UseStateForUnknown to:
- ip_pool
- name
- primary_interface

for network_interfaces and child_virtual_networks.  This suppresses and
spurious detection of changes in configuration.

We've added a check on network-interfaces to the Instance Update
isApiUpdateNeeded function since it is now safe to do so.