Metal 3: Fix to instance getStateInterfaces
In this PR we change getStateInterfaces() to return
[]NetworkInterfacesValue built from instance.interfaces as opposed to
that from the plan if the "shape" of []NetworkInterfacesValue as built
from instance.containerDetails.server.instances.  This will hopefully
fix the remaining issues seen with the network_interfaces State returned
for Metal instances.

As part of these changes we've renamed the compareAPIPlanInfts()
function to compareServerInstanceInfs() and also renamed the arguments
and variables within it.