sdv2: Refactor render functions to live in non-test packages (#728)
This PR moves each sdkv2 `Render{Resource}Config` function into the
respective `_example.go` file for each resource type.

The `Render{Resource}Config` functions were originally placed in the
`_test` packages.
This was problematic as it meant we couldn't import them for use as a
source of a configuration for testing dependencies.

Furthermore, they were unable to be used correctly as imports - the way
they were written did not correctly deduce the path of the `.tf.tmpl`
file in the call to `RenderExample` when called from an external
package. The syscall that `os.Read` used would return an `ENOENT` (Error
No Entry).

We switch to using this pattern for supporting `RenderExample` when
these functions are called from a different package:
```
	// Get the directory where this source file is located
	_, filename, _, ok := runtime.Caller(0)
	if !ok {
		return "", fmt.Errorf("unable to get current file path")
	}
	dir := filepath.Dir(filename)
	templatePath := filepath.Join(dir, "{resource}.tf.tmpl")

	return testhelpers.RenderExample(
		t,
		templatePath,
		"Name",
		defaults["Name"],
	)
```
This allows the `Render{Resource}Config` functions to be called from a
separate package correctly, so these can be used to generate
configurations for testing which support overrides.

Furthermore:
- Tidy up inconsistent code patterns in the `Render{Resource}Config`
functions.
- Fix naming of these functions (some of them were formatted like
`RenderMorpheus{Resource}Config` or `RenderHpeMorpheus{Resource}Config`.
- Change function signatures to accept a `name` parameter if the
resoruce requires a name (we could have this entirely handled by
overrides, but most of these render functions used a `name` parameter,
so I've stuck to using that if the resource requires a name).